requirements.txt
Flask
flask_bootstrap
flask_wtf
Redis
psycopg2-binary

